Implement second part of #33 : solo managers

+++ src/InsaLan/TournamentBundle/Controller/ManagerController.php
@@ -22,11 +22,13 @@ use Payum\Offline\PaymentFactory as OfflinePaymentFactory;
 use Payum\Paypal\ExpressCheckout\Nvp\Api;
 
 use InsaLan\TournamentBundle\Form\SetManagerName;
+use InsaLan\TournamentBundle\Form\SetPlayerName;
 use InsaLan\TournamentBundle\Form\TeamLoginType;
 use InsaLan\TournamentBundle\Exception\ControllerException;
 
 use InsaLan\TournamentBundle\Entity\Manager;
 use InsaLan\TournamentBundle\Entity\Participant;
+use InsaLan\TournamentBundle\Entity\Player;
 use InsaLan\TournamentBundle\Entity\Tournament;
 use InsaLan\TournamentBundle\Entity\Team;
 use InsaLan\TournamentBundle\Entity;
@@ -87,17 +89,78 @@ class ManagerController extends Controller
             $em->persist($manager);
             $em->flush();
 
-            return $this->redirect(
-                $this->generateUrl('insalan_tournament_manager_jointeamwithpassword', array('tournament' => $tournament->getId()))
-            );
+            if($tournament->getParticipantType() === "team")
+                return $this->redirect(
+                    $this->generateUrl('insalan_tournament_manager_jointeamwithpassword', array('tournament' => $tournament->getId()))
+                );
+            else // solo tournaments
+                return $this->redirect(
+                    $this->generateUrl('insalan_tournament_manager_joinsoloplayer', array('tournament' => $tournament->getId()))
+                );
         }
 
         return array('form' => $form->createView(), 'selectedGame' => $tournament->getType(), 'tournamentId' => $tournament->getId());
     }
 
     /**
+     * Allow a manager to join a player in a solo tournament
+     * @Route("/{tournament}/user/joinplayer")
+     * @Template()
+     */
+    public function joinSoloPlayerAction(Request $request, Entity\Tournament $tournament)
+    {
+        $em = $this->getDoctrine()->getManager();
+        $usr = $this->get('security.context')->getToken()->getUser();
+
+        // handle only solo tournaments
+        if($tournament->getParticipantType() !== "player")
+            throw new ControllerException("Joueurs solo non acceptées dans ce tournois");
+
+        // check if there is already a pending manager for this user and tournament
+        $manager = $em->getRepository('InsaLanTournamentBundle:Manager')
+            ->findOneByUserAndPendingTournament($usr, $tournament);
+
+        if($manager === null)
+            return $this->redirect($this->generateUrl('insalan_tournament_manager_setname', array('tournament' => $tournament->getId())));
+
+        $form_player = new Player();
+        $form = $this->createForm(new SetPlayerName(), $form_player); // fill player gamename
+        $form->handleRequest($request);
+
+        $error_details = null;
+        if ($form->isValid()) {
+            try {
+                // find the targeted player related to the manager
+                $player = $em
+                    ->getRepository('InsaLanTournamentBundle:Player')
+                    ->findOneBy(array(
+                        'gameName' => $form_player->getGameName(),
+                        'tournament' => $tournament));
+
+                if ($player === null)
+                    throw new ControllerException("Joueur introuvable !");
+                if ($player->getManager() != null)
+                    throw new ControllerException("Le joueur possède déjà un manager !");
+
+                $manager->setParticipant($player);
+                $player->setManager($manager);
+                $em->persist($manager);
+                $em->persist($player);
+                $em->flush();
+
+                return $this->redirect($this->generateUrl('insalan_tournament_manager_pay', array('tournament' => $tournament->getId())));
+
+            } catch (ControllerException $e) {
+                $error_details = $e->getMessage();
+            }
+
+        }
+        return array('tournament' => $tournament, 'user' => $usr, 'manager' => $manager, 'error' => $error_details, 'form' => $form->createView());
+    }
+
+    /**
      * Allow a new manager to join a team with name and password
-     * @Route("/{tournament}/user/setname")
+     * @Route("/{tournament}/user/jointeam")
      * @Template()
      */
     public function joinTeamWithPasswordAction(Request $request, Entity\Tournament $tournament)
@@ -300,9 +363,10 @@ class ManagerController extends Controller
 
         return array('tournament' => $tournament, 'user' => $usr, 'manager' => $manager);
     }
-    
+
     /**
      * Allow a manager to drop a pending tournament registration if not managed by team
+     * TODO add flashbag confirmation
      * @Route("/{tournament}/user/leave")
      */
     public function leaveAction(Entity\Tournament $tournament) {
@@ -316,6 +380,17 @@ class ManagerController extends Controller
         if($manager->getTournament()->getParticipantType() !== "player")
             throw new ControllerException("Not Allowed"); // must be a player only tournament
 
+        // not allowed if he paid something
+        if(!$tournament->isFree() && $manager->getPaymentDone()){
+            $this->get('session')->getFlashBag()->add('error', "Vous avez payé votre place, merci de contacter l'InsaLan si vous souhaitez vous désister.");
+            return $this->redirect($this->generateUrl('insalan_tournament_user_index'));
+        }
+        // not allowed either if registration are closed
+        if(!$tournament->isOpenedNow())
+            return $this->redirect($this->generateUrl('insalan_tournament_user_index'));
+
+        $manager->getParticipant()->setManager(null);
+        $manager->setParticipant(null);
         $em->remove($manager);
         $em->flush();
 
@@ -324,8 +399,8 @@ class ManagerController extends Controller
 
     /**
      * Allow a manager to drop a pending tournament registration managed by teams
+     * TODO add flashbag confirmation
      * @Route("/user/leave/team/{teamId}")
-     * @Template()
      */
     public function leaveTeamAction($teamId) {
         $em = $this->getDoctrine()->getManager();
@@ -353,7 +428,7 @@ class ManagerController extends Controller
         }
         // not allowed either if registration are closed
         if(!$team->getTournament()->isOpenedNow())
-            return $this->redirect($this->generateUrl('insalan_tournament_user_index')); 
+            return $this->redirect($this->generateUrl('insalan_tournament_user_index'));
 
         $manager->setParticipant(null);
         $team->setManager(null);
@@ -361,6 +436,7 @@ class ManagerController extends Controller
         $em->persist($team);
         $em->remove($manager);
         $em->flush();
+
         return $this->redirect($this->generateUrl('insalan_tournament_user_index'));
     }
